Rename actus-localstorage to persist and move to the main package
