+++ packages/actus-default-actions/package.json
@@ -3,6 +3,5 @@
   "version": "0.1.0-alpha.2",
   "repository": "https://github.com/EvgenyOrekhov/actus",
   "license": "MIT",
-  "author": "Evgeny Orekhov",
-  "files": []
+  "author": "Evgeny Orekhov"
 }
