+++ packages/actus/index.js
@@ -1,4 +1,5 @@
 import actus from "./src/init.js";
 import defaultActions from "./src/plugins/defaultActions/index.js";
+import persist from "./src/plugins/persist/index.js";
 
-export { actus, defaultActions };
+export { actus, defaultActions, persist };
